Optimize denormal check

+++ r4300/x86/gcop1_helpers.h
@@ -5,7 +5,7 @@ void gencheck_float_input_valid();
 void gencheck_float_output_valid();
 void gencheck_float_conversion_valid();
 
-extern const unsigned int largest_denormal_float;
-extern const unsigned long long largest_denormal_double;
+extern const float largest_denormal_float;
+extern const double largest_denormal_double;
 
 #endif // GCOP1_HELPERS_H
